requirements.txt
colour==0.1.5
faiss-gpu==1.6.5
hydra-core==1.0.6
numpy==1.19.4
pandas==1.1.5
tensorboard==2.4.1
torch==1.7.1
torchinfo==0.0.8
tqdm==4.51.0
wsiprocess==0.8